通知型プロパティに各所を変更した

+++ Client/Models/ChatModel.cs
@@ -13,13 +13,28 @@ namespace ChatApp.Client.Models
 {
     public class ChatModel
     {
-        public List<PostModel> PostModels { get; private set; }
+        private PostModel[] _postModels;
+        public PostModel[] PostModels
+        { 
+            get => _postModels; 
+            private set
+            { 
+                _postModels = value;
+                PostsChanged?.Invoke(this, _postModels);
+            }
+        }
 
+        public UserInformation[] _roomParticipants;
         // 入室中のユーザー情報
-        public List<UserInformation> RoomParticipants { get; private set; }
-
-        // すべてのユーザー
-        public List<UserInformation> AllUser { get; private set; }
+        public UserInformation[] RoomParticipants 
+        { 
+            get => _roomParticipants;
+            private set
+            {
+                _roomParticipants = value;
+                RoomParticipantsChanged?.Invoke(this, _roomParticipants);
+            }
+        }
 
         public Guid RoomId { get; private set; }
 
@@ -31,19 +46,25 @@ namespace ChatApp.Client.Models
 
         private readonly IHttpClientFactory _httpClientFactory;
 
-        private readonly HubConnection _hubConnection;
+        private readonly HubUtility _hubUtility;
 
+        private HubConnection _hubConnection;
+
+        private IndexModel _indexModel;
         /// <summary>
         /// コンストラクタ
         /// </summary>
-        /// <param name="httpClientFactory"></param>
-        /// <param name="hubConnection"></param>
-        /// <param name="roomId"></param>
-        public ChatModel(IHttpClientFactory httpClientFactory, HubUtility HubUtility)
+        public ChatModel(IHttpClientFactory httpClientFactory, HubUtility hubUtility, IndexModel indexModel)
         {
             _httpClientFactory = httpClientFactory;
-            _hubConnection = hubConnection;
+            _hubUtility = hubUtility;
+            _indexModel = indexModel;
+        }
 
+        public async Task Initialize(Guid roomId)
+        {
+            await UpdateRoomModelAsync(roomId);
+            _hubConnection = _hubUtility.CreateHubConnection();
             _hubConnection.On<Message>(SignalRMehod.ReceiveMessage, async (message) =>
             {
                 if (message.RoomId != RoomId)
@@ -72,10 +93,7 @@ namespace ChatApp.Client.Models
                             UpdateDate = threadMessage.UpdateDate
                         }).ToList()
                 };
-
-                PostModels.Add(postModel);
-
-                PostsChanged(this, PostModels.ToArray());
+                PostModels = PostModels.Append(postModel).ToArray();
             });
 
             _hubConnection.On<ThreadMessage>(SignalRMehod.SendThreadMessage, (message) =>
@@ -113,30 +131,15 @@ namespace ChatApp.Client.Models
 
                 ThreadsChanged(this, threads.ToArray());
             });
-        }
 
-        public async Task Initialize(Guid roomId)
-        {
-            await UpdateRoomModelAsync(roomId);
             await _hubConnection.StartAsync();
         }
 
         public async Task UpdateRoomModelAsync(Guid roomId)
         {
             RoomId = roomId;
-            await GetAllUserAsync();
-            await GetAllRoomParticipantsAsync();
+            await GetAllRoomParticipantsAsync(roomId);
             await GetRoomPostsAsync();
-        }
-
-        /// <summary>
-        /// サービスに登録しているすべてのユーザーを取得する
-        /// </summary>
-        /// <returns></returns>
-        public async Task GetAllUserAsync()
-        {
-            var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
-            AllUser = await client.GetFromJsonAsync<List<UserInformation>>("User");
         }
 
         /// <summary>
@@ -144,12 +147,12 @@ namespace ChatApp.Client.Models
         /// </summary>
         /// <param name="roomId"></param>
         /// <returns></returns>
-        public async Task GetAllRoomParticipantsAsync()
+        public async Task GetAllRoomParticipantsAsync(Guid roomId)
         {
             var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
-            var roomDetail = await client.GetFromJsonAsync<RoomDetail>($"Room/{RoomId}");
+            var roomDetail = await client.GetFromJsonAsync<RoomDetail>($"Room/{roomId}");
 
-            RoomParticipants = roomDetail.Users;
+            RoomParticipants = roomDetail.Users.ToArray();
 
             RoomParticipantsChanged?.Invoke(this, RoomParticipants.ToArray());
         }
@@ -198,9 +201,7 @@ namespace ChatApp.Client.Models
                 };
             }));
 
-            PostModels = postModelTask.ToList();
-
-            PostsChanged?.Invoke(this, PostModels.ToArray());
+            PostModels = postModelTask;
         }
 
         /// <summary>
@@ -212,9 +213,7 @@ namespace ChatApp.Client.Models
         {
             var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
             var response = await RoomUtility.AddUsersToRoom(RoomId, userEmails, client);
-            RoomParticipants = response.Users;
-
-            RoomParticipantsChanged?.Invoke(this, RoomParticipants.ToArray());
+            RoomParticipants = response.Users.ToArray();
         }
 
         /// <summary>
@@ -226,9 +225,7 @@ namespace ChatApp.Client.Models
         {
             var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
             var response = await RoomUtility.DeleteUserFromRoom(RoomId, userEmails, client);
-            RoomParticipants = response.Users;
-
-            RoomParticipantsChanged?.Invoke(this, RoomParticipants.ToArray());
+            RoomParticipants = response.Users.ToArray();
         }
 
         /// <summary>
@@ -253,9 +250,7 @@ namespace ChatApp.Client.Models
 
             var deletedMessage = PostModels.FirstOrDefault(postModel => postModel.PostId == postId);
 
-            PostModels.Remove(deletedMessage);
-
-            PostsChanged?.Invoke(this, PostModels.ToArray());
+            PostModels = PostModels.Where(postModel => postModel.PostId == postId).ToArray();
         }
 
         /// <summary>
@@ -267,6 +262,7 @@ namespace ChatApp.Client.Models
         public async Task ChangeMessageAsync(Guid postId, string newMessage)
         {
             var post = PostModels.FirstOrDefault(post => post.PostId == postId);
+            if (post == null) return;
 
             var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
             var result = await client.PutAsJsonAsync($"Post/{postId}", new ChatPostUpdateRequest() { Text = newMessage });
@@ -282,6 +278,7 @@ namespace ChatApp.Client.Models
         public async Task GetThreadMessageAsync(Guid postId)
         {
             var threads = PostModels.FirstOrDefault(post => post.PostId == postId)?.ThreadModels;
+            if(threads == null) return;
 
             var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
             var result = await client.GetFromJsonAsync<List<ThreadMessage>>("Thread/Post/" + postId.ToString());
@@ -324,6 +321,7 @@ namespace ChatApp.Client.Models
         public async Task DeleteThreadMessageAsync(Guid postId, Guid threadId)
         {
             var threads = PostModels.FirstOrDefault(post => post.PostId == postId).ThreadModels;
+            if (threads == null) return;
 
             var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
             await client.DeleteAsync($"Thread/{threadId}");
@@ -339,12 +337,14 @@ namespace ChatApp.Client.Models
             var thread = PostModels.FirstOrDefault(post => post.PostId == message.PostId).ThreadModels
                 .FirstOrDefault(thread => thread.ThreadId == message.ThreadId);
 
+            if (thread == null) return;
+
             var client = _httpClientFactory.CreateClient("ChatApp.ServerAPI");
             await client.PutAsJsonAsync("Thread", message);
 
             thread.MessageContext = message.MessageContext;
 
-            ThreadsChanged(this, new ThreadModel[] { thread });
+            ThreadsChanged?.Invoke(this, new ThreadModel[] { thread });
         }
     }
 }
